Task 7.3 Forum class moved to forumuser package
